requirements.txt
# Core dependencies
pyyaml>=6.0
google-generativeai>=0.3.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# UI Framework (optional - for Streamlit)
streamlit>=1.28.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Testing and validation
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dateutil>=2.8.2
requests>=2.31.0
python-dotenv>=1.0.0

# Financial data
yfinance>=0.2.0
alpha-vantage>=2.3.1
feedparser>=6.0.10
beautifulsoup4>=4.12.0

# Text processing for retrieval
sentence-transformers>=2.2.0
scikit-learn>=1.3.0